More work converting between Plutus and Wallet types
