+++ lib/shelley/cardano-wallet.cabal
@@ -307,6 +307,7 @@ test-suite integration
       shelley-integration-test.hs
   other-modules:
       Cardano.Wallet.Shelley.Faucet
+      Cardano.Wallet.Shelley.Integration
 
 benchmark restore
   default-language:
